% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fcns1.R
\name{sim_CDR}
\alias{sim_CDR}
\title{sim_CDR}
\usage{
sim_CDR(n, time, G, W_ind, effect_ind, vars, effects, main_effects)
}
\arguments{
\item{n}{sample size}

\item{time}{time intervals, as in, 0:5}

\item{G}{correlation matrix}

\item{vars}{patient characteristics}

\item{effects}{effects placed on the patient characteristics}

\item{main_effects}{specifying for Y_0, rate and rate_A (treatment reduction of rate)}
}
\description{
Simulating Alzheimer's data
}
